:truck: [#1056] -- move EIDAS auth plugin into eherkenning package
The implementation is nearly identical and the eidas app was already
importing from the eherkenning app. It makes more sense to keep
those together in the codebase.
